requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP and networking
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Parsing and data processing
beautifulsoup4==4.12.2
lxml==4.9.4
python-nmap==0.7.1
dnspython==2.4.2

# Reporting
jinja2==3.1.2
markdown==3.5.1
reportlab==4.0.7
weasyprint==60.1

# API and Dashboard
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Logging and monitoring
loguru==0.7.2
python-json-logger==2.0.7

# Security tools integration
python-whois==0.8.0
shodan==1.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0
